tomi_sol/requirements.txt
openai
flask
requests
beautifulsoup4
python-dotenv